surveyLocations.R
This plots the location of survey takes at the national scale and Oklahoma state scale. The Qualtrics API was working until I changed the API key one time, so there is an error with that section right now. An email has been sent into why it's not working anymore. Still works if you download the Qualtrics data from the website.