Add files via upload
This code will upload the newest data from qualtrics and combine the mturk and online surveys. Along with trimming the data down to the columns we need and the survey takers that passed the attention questions. The qualtrics_survey.csv is the result of the code and the csv that can be used for the next time to reduce the amount of people that you have to pass or fail on the override.